Added wrapper to ensure compatibility with attestation.id main branch
